src/raspi/requirements.txt
sanic>=0.7.0 #Webframework https://github.com/huge-success/sanic
protobuf>=3.3.0 #middleware
pyzmq>=17.1.0 #middleware
pyserial>=3.4 #interface raspi and tiny